Implement notification handler for pull requests
To handle the pull request, the merge() method of git has to be changed;
this commit introduces the changes. Pull requests will work only with
compare branches that are on the same repository as the base branch, so
forks are currently not supported.

Also, tests for the new type are introduced.