Add verification of github signature
Github webhooks can be configured such that the request header contains
a signature of the payload, a secret is used as the key. This secret can
be configured in the webhook settings.

GCI verifies this signature if the secret is also set in the gnampfile.
If the secret is not set or empty, GCI does not verify the signature.

Please note that the tests do not concern the signature verification.