Implement first working prototype of build worker
This commit introduces a first working version of the build worker. The
worker must be started with its id and the main server's remote address
as params. It will afterwards connect to the main server and request the
git credentials and the build script.
After executing the tests, the worker will send the test results. While
executing the tests, the worker will stream their command line output.