Change Create() and Add Wait()
Create() now creates a container with a TTY. This is only a solution for
the time beeing as it eases the handling of the output stream. The tests
are modified accordingly.

Wait() waits for the container to be in a not-running state and converts
the result into the internal BuildResult, which can represent either a
success, an error in one of the three stages or a general error.